4 spec done done. 4/11 (Last day I'm here except for maybe this sunday)
Agenda:
We need to fully code a sample auto, finish up Tele-op, and finalize any other changes that come to attention in the following meeting(s). The only reason I'm writing this so fancily is so I will know what I need to do. Aside from that, the robot should be functioning decently*

*be honest now